public AGent

+++ utils/interceptor.js
@@ -2,10 +2,18 @@ import axios from "axios";
 
 axios.interceptors.request.use(
     async (config) => {
-        let token = localStorage.getItem("proxy_token");
-        config.headers['proxy_auth_token'] = token;
-        if (process.env.NEXT_PUBLIC_ENV === 'local')
-            config.headers['Authorization'] = localStorage.getItem("local_token");
+        if(config.url?.includes("publicAgent")){
+            const token = config.url?.includes("login") ? localStorage.getItem("publicAgentProxyToken"): null
+            let AuthToken = localStorage.getItem("AgentToken")
+            config.headers['Authorization'] = !config.url?.includes("login") ? AuthToken : null;
+            config.headers['proxy_auth_token'] = token;
+        }
+        else{
+            let token = localStorage.getItem("proxy_token");
+            config.headers['proxy_auth_token'] = token;
+            if (process.env.NEXT_PUBLIC_ENV === 'local')
+                config.headers['Authorization'] = localStorage.getItem("local_token");
+        }
         return config;
     },
     (error) => {
@@ -19,9 +27,10 @@ axios.interceptors.response.use(
         return response;
     },
     async function (error) {
-        if (error?.response?.status === 401) {
+        console.log(error);
+        if (error?.response?.status === 401 && !error?.config?.url?.includes("publicAgent")) {
             localStorage.clear();
-            if(window.location.href!='/login')localStorage.setItem("previous_url", window.location.href);           
+            localStorage.setItem("previous_url", window.location.href);           
             window.location.href = "/login";
         }
         return Promise.reject(error);
